Introduce build flag RELEASE_HIDDEN_API_EXPORTABLE_STUBS
The new build flag determines whether the hiddenapi flags depend on the
"everything" stubs (i.e. the stubs that include all flagged apis) or the
"exportable" stubs (i.e. the stubs that only include enabled flagged
apis). The flag is defaulted to false, thus the hiddenapi will depend on
the "everything" stubs by default.

This is a temporary solution until the metalava fully supports flagged
apis handling, and will be removed once the metalava feature
implementation is complete.

Test: m nothing
Bug: 315027929
Change-Id: Ib08a830891634e07a8b22e6f163cefc6c13ff37a