Set default value of RELEASE_DEPRECATE_VNDK as True
RELEASE_DEPRECATE_VNDK flag was created to deprecate VNDK as a release
flag. VNDK deprecation is completed, so it should be deprecated by
default, and should be enabled unless there is any other issue. This
change set default value of RELEASE_DEPRECATE_VNDK as True, so it can be
applied to all release flags by default.

Bug: 329737440
Test: AOSP CF build succeeded
Change-Id: I3bf87b467a615d3a2e3dc2886ca5fcead0634c12