Advance trunk_staging to sqlite 3.44.3 on main
Tested by building and installing an image on a phone and verifying
that the phone is running 3.44.3.

Build flags are normally aosp-first but multiple commits directly into
the flag file on main have introduced merge conflicts.  After this
commit has been made, the aosp files will be updated.

Bug: 322547687

Ignore-AOSP-First: resolving a merge conflict.

Change-Id: Icc63e3fa618fcf5370efabb41af12851906e0a51