Add a build flag for storage areas
This adds a build flag for storage areas, a new feature that is
being developed on aosp/main. We need this build flag in order to
gate some new SELinux policies.
However, since it seems trunk_staging.scl is developed here on
internal main and our CLs in aosp/main are reporting merge conflicts,
we are submitting the build flag addition in this repo and will
pick it onto our aosp code.

Bug: 325129836
Test: m
Ignore-AOSP-First: merge conflict resolution
Change-Id: I54cfa154f8a0a10a1e06763c3243b92f95c4b1cd