Advance sqlite 3.44.3 to trunk on main
Tested by building and installing an image on a phone and verifying
that the phone is running 3.44.3.

Build flags are normally aosp-first but multiple commits directly into
the flag file on main have introduced merge conflicts.  After this
commit has been made, the aosp files will be updated.

Bug: 322547687

Ignore-AOSP-First: resolving a merge conflict.
Change-Id: Ieac131c5692559551a6912bf4deb4584155b56ba