Removing flag com.android.trunk_stable_workflow_testing.test_migrate_flag from RELEASE_CONFIG_TRUNKFOOD_STAGING.
This commit was automatically created by advancing a workflow in Gantry. Manual edits to this file are forbidden. If you need to rollback, please visit  to begin a rollback.

POSTSUBMIT ERRORS: If this change breaks postsubmits, please revert it and attach the associated bugs to the reverted CL's description.

Change-Id: I5e1281d23906992ee0e3c37e47b90839f63b6839
Ignore-AOSP-First: AConfig change for future AOSP release