Removing flag com.example.android.aconfig.demo.flags.yet_another_mendel_gantry_disintegration_flag from RELEASE_CONFIG_TRUNKFOOD_STAGING.
This commit was automatically created by advancing a workflow in Gantry. Manual edits to this file are forbidden. If you need to rollback, please visit  to begin a rollback.

POSTSUBMIT ERRORS: If this change breaks postsubmits, please revert it and attach the associated bugs to the reverted CL's description.

Change-Id: I22228d144972c6d875e2c0aa9e54fad8295db5cf
Ignore-AOSP-First: AConfig change for future AOSP release